# docs/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AnyChat 即时通讯系统的网关 API 服务，提供用户认证、用户管理等功能的 HTTP 接口。\n所有需要认证的接口都需要在 Header 中携带 Authorization: Bearer \u003ctoken\u003e",
        "title": "AnyChat Gateway API",
        "contact": {
            "name": "AnyChat API Support",
            "url": "https://github.com/yzhgit/anchat_server",
            "email": "support@anychat.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户加入的所有群组列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "获取我的群组列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "最后更新时间戳（增量同步）",
                        "name": "lastUpdateTime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_anychat_server_internal_group_dto.GroupListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的群组，至少需要2个成员（包括创建者）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "创建群组",
                "parameters": [
                    {
                        "description": "群组信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_internal_group_dto.CreateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_anychat_server_internal_group_dto.GroupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定群组的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "获取群组信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_anychat_server_internal_group_dto.GroupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "群组不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新群组的名称、头像、公告等信息（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "更新群信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_internal_group_dto.UpdateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "解散群组（仅群主可操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "解散群组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "解散成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/groups/{id}/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "申请加入指定群组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "加入群组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "申请信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_internal_group_dto.JoinGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "申请成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_anychat_server_internal_group_dto.JoinGroupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/groups/{id}/members/{userId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从群组中移除成员（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "移除成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要移除的用户ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "移除成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/github_com_anychat_server_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_anychat_server_internal_group_dto.GroupListResponse": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_anychat_server_internal_group_dto.GroupResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 5
                },
                "updateTime": {
                    "type": "integer",
                    "example": 1609459200
                }
            }
        },
...
        "github_com_anychat_server_internal_group_dto.UpdateGroupRequest": {
            "type": "object",
            "properties": {
                "announcement": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "群公告内容"
                },
                "avatar": {
                    "type": "string",
                    "example": "https://example.com/new-avatar.jpg"
                },
                "joinVerify": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "新群名称"
                }
            }
        },
...
        "github_com_anychat_server_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}